Public Documentation

Documentation for SpatialEcology.jl's public interface.

Contents

Public Documentation

Contents

Index

Function documentation

Index

DataFrames.aggregate

RandomBooleanMatrices.matrixrandomizer

StatsAPI.pairwise

DataFramesMeta.@with

Function documentation

Types

Missing docstring.

Missing docstring for Assemblage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ComMatrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SiteData. Check Documenter's build log for details.

Getting attributes

Missing docstring.

Missing docstring for nspecies. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nsites. Check Documenter's build log for details.

Missing docstring.

Missing docstring for occupancy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for richness. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nrecords. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sitenames. Check Documenter's build log for details.

Missing docstring.

Missing docstring for speciesnames. Check Documenter's build log for details.

Missing docstring.

Missing docstring for coordinates. Check Documenter's build log for details.

Missing docstring.

Missing docstring for occurring. Check Documenter's build log for details.

Missing docstring.

Missing docstring for noccurring. Check Documenter's build log for details.

Missing docstring.

Missing docstring for occupied. Check Documenter's build log for details.

Missing docstring.

Missing docstring for noccupied. Check Documenter's build log for details.

Missing docstring.

Missing docstring for occurrences. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cooccurring. Check Documenter's build log for details.

Extracting and setting data

Missing docstring.

Missing docstring for traits. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sitestats. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sitestatnames. Check Documenter's build log for details.

Missing docstring.

Missing docstring for traitnames. Check Documenter's build log for details.

Missing docstring.

Missing docstring for commatrix. Check Documenter's build log for details.

DataFramesMeta.@with — Macro

@with(d, expr)

@with allows DataFrame columns keys to be referenced as symbols.

Arguments

d : an AbstractDataFrame type

expr : the expression to evaluate in d

Details

@with works by parsing the expression body for all columns indicated by symbols (e.g. :colA). Then, a function is created that wraps the body and passes the columns as function arguments. This function is then called. Operations are efficient because:

A pseudo-anonymous function is defined, so types are stable.

Columns are passed as references, eliminating DataFrame indexing.

The following

@with(d, :a .+ :b .+ 1)

becomes

tempfun(a, b) = a .+ b .+ 1
tempfun(d[!, :a], d[!, :b])

If an expression is wrapped in ^(expr), expr gets passed through untouched. If an expression is wrapped in $(expr), the column is referenced by the variable expr rather than a symbol.

If the expression provide to @with begins with @byrow, the function created by the @with block is broadcasted along the columns of the data frame.

Examples

julia> using DataFramesMeta

julia> y = 3;

julia> df = DataFrame(x = 1:3, y = [2, 1, 2]);

julia> x = [2, 1, 0];

julia> @with(df, :y .+ 1)
3-element Vector{Int64}:
 3
 2
 3

julia> @with(df, :x + x)
3-element Vector{Int64}:
 3
 3
 3

julia> @with df begin
            res = 0.0
            for i in 1:length(:x)
                res += :x[i] * :y[i]
            end
            res
        end
10.0

julia> @with(df, df[:x .> 1, ^(:y)]) # The ^ means leave the :y alone
2-element Vector{Int64}:
 1
 2

julia> colref = :x;

julia> @with(df, :y + $colref) # Equivalent to df[!, :y] + df[!, colref]
3-element Vector{Int64}:
 3
 3
 5

julia> @with df @byrow :x * :y
3-element Vector{Int64}:
 2
 2
 6

Note

@with creates a function, so the scope within @with is a local scope. Variables in the parent can be read. Writing to variables in the parent scope differs depending on the type of scope of the parent. If the parent scope is a global scope, then a variable cannot be assigned without using the global keyword. If the parent scope is a local scope (inside a function or let block for example), the global keyword is not needed to assign to that parent scope.

Missing docstring.

Missing docstring for @traits. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @sitestats. Check Documenter's build log for details.

Missing docstring.

Missing docstring for addtraits!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for addsitestats!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sitetotals. Check Documenter's build log for details.

Missing docstring.

Missing docstring for speciestotals. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getspecies. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getsite. Check Documenter's build log for details.

Setting data

Missing docstring.

Missing docstring for coordstype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for xcells. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ycells. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cells. Check Documenter's build log for details.

Missing docstring.

Missing docstring for xmin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for xmax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ymin. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ymax. Check Documenter's build log for details.

Missing docstring.

Missing docstring for xrange. Check Documenter's build log for details.

Missing docstring.

Missing docstring for yrange. Check Documenter's build log for details.

Missing docstring.

Missing docstring for xcellsize. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ycellsize. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cellsize. Check Documenter's build log for details.

Missing docstring.

Missing docstring for boundingbox. Check Documenter's build log for details.

Grouping

Missing docstring.

Missing docstring for groupspecies. Check Documenter's build log for details.

Missing docstring.

Missing docstring for groupsites. Check Documenter's build log for details.

DataFrames.aggregate — Function

aggregate(object, grid [, fun])

Aggregate object (either an Assemblage or Locations type) to grid. If object is an Assemblage{PointData} this will grid all points and return an Assemblage{GridData}. grid can be a GridTopology or a single Integer signifying the aggregation factor for already gridded data, the cellsize for point data. fun is an optional function specifying how to lump occurrences. If not specified the default function is any for Boolean Assemblages and sum for Integer ones.

Randomization

RandomBooleanMatrices.matrixrandomizer — Function

matrixrandomizer(m [,rng]; method = curveball)

Create a matrix generator function that will return a random boolean matrix every time it is called, maintaining row and column sums. Non-boolean input matrix are interpreted as boolean, where values != 0 are true.

Examples

``` m = rand(0:4, 5, 6) rmg = matrixrandomizer(m)

random1 = rand(rmg) random2 = rand(rmg) ``

Missing docstring.

Missing docstring for matrixrandomizations. Check Documenter's build log for details.

Utilities

Missing docstring.

Missing docstring for dispersionfield. Check Documenter's build log for details.

StatsAPI.pairwise — Function

pairwise(f, x[, y];
         symmetric::Bool=false, skipmissing::Symbol=:none)

Return a matrix holding the result of applying f to all possible pairs of entries in iterators x and y. Rows correspond to entries in x and columns to entries in y. If y is omitted then a square matrix crossing x with itself is returned.

As a special case, if f is cor, diagonal cells for which entries from x and y are identical (according to ===) are set to one even in the presence missing, NaN or Inf entries.

Keyword arguments

symmetric::Bool=false: If true, f is only called to compute for the lower triangle of the matrix, and these values are copied to fill the upper triangle. Only allowed when y is omitted. Defaults to true when f is cor or cov.

skipmissing::Symbol=:none: If :none (the default), missing values in inputs are passed to f without any modification. Use :pairwise to skip entries with a missing value in either of the two vectors passed to f for a given pair of vectors in x and y. Use :listwise to skip entries with a missing value in any of the vectors in x or y; note that this might drop a large part of entries. Only allowed when entries in x and y are vectors.

Examples

julia> using StatsBase, Statistics

julia> x = [1 3 7
            2 5 6
            3 8 4
            4 6 2];

julia> pairwise(cor, eachcol(x))
3×3 Matrix{Float64}:
  1.0        0.744208  -0.989778
  0.744208   1.0       -0.68605
 -0.989778  -0.68605    1.0

julia> y = [1 3 missing
            2 5 6
            3 missing 2
            4 6 2];

julia> pairwise(cor, eachcol(y), skipmissing=:pairwise)
3×3 Matrix{Float64}:
  1.0        0.928571  -0.866025
  0.928571   1.0       -1.0
 -0.866025  -1.0        1.0

pairwise(metric::PreMetric, a::AbstractMatrix, b::AbstractMatrix=a; dims)

Compute distances between each pair of rows (if dims=1) or columns (if dims=2) in a and b according to distance metric. If a single matrix a is provided, compute distances between its rows or columns.

a and b must have the same numbers of columns if dims=1, or of rows if dims=2.

pairwise(metric::PreMetric, a, b=a)

Compute distances between each element of collection a and each element of collection b according to distance metric. If a single iterable a is provided, compute distances between its elements.

Missing docstring.

Missing docstring for asquantiles. Check Documenter's build log for details.

Missing docstring.

Missing docstring for asquantiles!. Check Documenter's build log for details.