Constructing SpatialEcology objects

Constructing SpatialEcology objects

Index

API

Put texts here

Index

SpatialEcology.Assemblage

SpatialEcology.ComMatrix

SpatialEcology.SiteData

API

SpatialEcology.Assemblage — Type

Assemblage{D<:Real, P<:Locations}

SpatialEcology.ComMatrix — Type

Commatrix

A type that holds a species-by-sites matrix of either presence data or abundances, along with the names of species and sites

SpatialEcology.SiteData — Type

SiteData

A type to hold spatial data on a set of sites